+++ components/accelerator/nvidia/clock-speed/clock_speed.go
@@ -6,7 +6,7 @@ import (
 	"github.com/NVIDIA/go-nvml/pkg/nvml"
 
 	"github.com/leptonai/gpud/pkg/nvidia-query/nvml/device"
-	nvmlerrors "github.com/leptonai/gpud/pkg/nvidia/error"
+	nvmlerrors "github.com/leptonai/gpud/pkg/nvidia/errors"
 )
 
 // ClockSpeed represents the data from the nvmlDeviceGetClockInfo API.
