+++ components/accelerator/nvidia/nvlink/nvlink_test.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"testing"
 
-	nvmlerrors "github.com/leptonai/gpud/pkg/nvidia/error"
+	nvmlerrors "github.com/leptonai/gpud/pkg/nvidia/errors"
 
 	"github.com/NVIDIA/go-nvml/pkg/nvml"
 	"github.com/stretchr/testify/assert"
