+++ components/accelerator/nvidia/remapped-rows/remapped_rows.go
@@ -7,7 +7,7 @@ import (
 	"github.com/leptonai/gpud/pkg/nvidia-query/nvml/device"
 
 	"github.com/leptonai/gpud/pkg/log"
-	nvmlerrors "github.com/leptonai/gpud/pkg/nvidia/error"
+	nvmlerrors "github.com/leptonai/gpud/pkg/nvidia/errors"
 )
 
 // RemappedRows represents the row remapping data.
