+++ pkg/nvidia/errors/error.go
