+++ pkg/nvidia/errors/error_test.go
